backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
google-cloud-storage
pandas
openpyxl
python-multipart
sqlalchemy
google-generativeai
